Calculate current NIFTY level and nearest option details
Added route to find current NIFTY level. Then we get all the instruments having NIFTY as underlying and expiring in current month, we filter it down to instruments expiring in currenty expiry that usually happens on Thursday). Now, from current week's expiry options we find out the nearest options to current NIFTY level and print details. These are the instruments we want tplay with!!